% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_death.R
\docType{data}
\name{covid_death}
\alias{covid_death}
\title{Confirmed Covid Deaths}
\format{
Data are structured as one observation per county.
Variables include:
\describe{
\item{county_fips}{County FIPS code (unique ID)}
\item{county_name}{County name}
\item{state}{State Name}
\item{state_fips}{State FIPS code (unique ID)}
\item{jan22}{Number of confirmed covid deaths on January 22,2020}
}
}
\source{
These data were combiled by Will Doyle and downloaded from:
\url{https://github.com/wdoyle42/ya_covid_dash}
Data were originally sourced from USA Facts
\url{https://usafacts.org/visualizations/coronavirus-covid-19-spread-map/}
}
\usage{
data(covid_death)
}
\description{
Loads clean version of USA Facts's confirmed covid death data into the environment.
}
\details{
These data track COVID-19 deaths by county from January 22,2020 to May 21, 2020
}
\keyword{datasets}
