% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djibril.R
\name{djibril}
\alias{djibril}
\title{Création d'histogramme, boite à moustaphes et résumé des statistiques descriptives}
\usage{
djibril(x)
}
\arguments{
\item{x}{variable numérique}
}
\description{
Création d'histogramme, boite à moustaphes et résumé des statistiques descriptives
}
