% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.R
\name{ds}
\alias{ds}
\title{Creates histogram, boxplot, and numerical summary of a numeric vector.}
\usage{
ds(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A data frame with mean, min, max, and median.
}
\description{
Creates histogram, boxplot, and numerical summary of a numeric vector.
}
