% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-surplus.R
\name{print.surplus}
\alias{print.surplus}
\title{print.surplus}
\usage{
\method{print}{surplus}(x, top = length(x$Ut))
}
\arguments{
\item{x}{a surplus class object}

\item{top}{number of rows to be printed}
}
\description{
print function for surplus
}
\examples{
print(surplus(c(1.2, 3.4), 4, 3, 30))

}
