\name{X}
\alias{X}

\title{
X function
}

\description{
The function performs input tables of the environment R
}

\usage{
X(x)
}

\arguments{
  \item{x}{
x is NULL
}
}

\details{

insert

X ()

select the desired table and press enter

observation: the mouse cursor should be in front of X ()
}

\value{
returns a data.frame
}

\author{
Emmanuel Arnhold <emmanuelarnhold@yahoo.com.br>
}


\seealso{
gds, dscor
}

\examples{
#X()
}

