% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.R
\name{ds}
\alias{ds}
\title{Descriptive statistics information}
\usage{
ds(x)
}
\arguments{
\item{x}{a numeric value input}
}
\value{
returns some descriptive statistics
}
\description{
Descriptive statistics information
}
\examples{
ds(rnorm(100))
}
